:tada: Load probabilities fixed between 0 and 1

+++ src/clover_gui/details/solar.py
@@ -983,7 +983,9 @@ class PVFrame(ttk.Frame):
 
         for pv_panel in pv_panels:
             self.panel_name_values[pv_panel.name] = ttk.StringVar(self, pv_panel.name)
-            self.panel_lifetimes[pv_panel.name] = ttk.DoubleVar(self, pv_panel.lifetime)
+            self.panel_lifetimes[pv_panel.name] = ttk.IntVar(
+                self, int(pv_panel.lifetime)
+            )
 
             # Panel orientation
             self.panel_tilt[pv_panel.name] = ttk.DoubleVar(
